+++ app/lib/widget/dialogs/quick_actions_dialog.dart
@@ -112,8 +112,8 @@ class _QuickActionsDialogState extends State<QuickActionsDialog> with Refena {
         ),
         ElevatedButton(
           style: ElevatedButton.styleFrom(
-            backgroundColor: Theme.of(context).buttonTheme.colorScheme!.primary,
-            foregroundColor: Theme.of(context).buttonTheme.colorScheme!.onPrimary,
+            backgroundColor: Theme.of(context).colorScheme.primary,
+            foregroundColor: Theme.of(context).colorScheme.onPrimary,
           ),
           onPressed: () {
             switch (_action) {