+++ app/assets/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - feat: ignore duplicate files when selected from file picker (@programmermager)
 - feat: add donation options (@Tienisto)
+- feat: add Yaru theme (@Tienisto)
 - feat(macos): add Command+W shortcut to close the window (@Q1CHENL)
 - fix: also show an OLED color mode option when dynamic colors are not supported by OS (@dhruvanbhalara)
 - fix(ios): fix permission error when picking directory (@Tienisto)